+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.0.4] - 28 December 2020
+
+* DeltaText line break bug fixed.
+
+## [0.0.3] - 28 December 2020
+
+* DeltaText style added.  
+* DeltaTextEditingController bug fixed.
+
 ## [0.0.2] - 27 December 2020
 
 * TextView added.